# api/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para registro e consulta de contratos de financiamento automotivo utilizando blockchain Hyperledger Besu e contratos inteligentes UUPS",
        "title": "MobX API - VFinance Registry",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "MobX API Support",
            "email": "support@mobx-api.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "3.0.0"
    },
    "host": "144.22.179.183:3000",
    "basePath": "/",
    "paths": {
...
        "/api/metadata/{hash}": {
            "get": {
                "description": "Busca metadados de um veículo pelo hash",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "metadata"
                ],
                "summary": "Get Metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hash dos metadados",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Metadados encontrados",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Hash inválido",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Metadados não encontrados",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atualiza metadados de um veículo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "metadata"
                ],
                "summary": "Update Metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hash dos metadados",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados do veículo",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Metadados atualizados com sucesso",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Armazena metadados de um veículo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "metadata"
                ],
                "summary": "Store Metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hash dos metadados",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados do veículo",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Metadados armazenados com sucesso",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove metadados de um veículo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "metadata"
                ],
                "summary": "Delete Metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hash dos metadados",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Metadados removidos com sucesso",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Hash inválido",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Digite \"Bearer\" seguido de um espaço e o token JWT",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}